adding a way to report transactions

+++ src/block.rs
@@ -12,7 +12,7 @@ const COST: u32 = 7;
 pub struct DataPoint {
   from: String,
   to: String,
-  amount: u64,
+  amount: f64,
 }
 
 impl DataPoint {
@@ -54,7 +54,7 @@ fn hash_with_cost(data: String) -> (String, u64) {
 }
 
 impl Block {
-  pub fn new(from: String, to: String, amount: u64, previous_hash: String) -> Block {
+  pub fn new(from: String, to: String, amount: f64, previous_hash: String) -> Block {
     let (summary, nonce) =
       hash_with_cost(from.clone() + &to + &previous_hash + amount.to_string().as_str());
 
@@ -74,7 +74,7 @@ impl Block {
       data: DataPoint {
         from: "NOONE".to_string(),
         to: "NOONE".to_string(),
-        amount: 0,
+        amount: 0.0,
       },
       previous_summary: "NONE".to_string(),
       nonce: 0,
